refactor: reparent highlighter specific types
