+++ src/Highlight/types.ts
@@ -29,3 +29,27 @@ export interface IHighlightCollection extends EventEmitter {
 export interface IHighlightCollectionAppModel {
 	readonly highlightCollection: IHighlightCollection;
 }
+
+export interface ColorDescription {
+	name: string;
+	val: string;
+}
+
+export enum MessageAction {
+	TOGGLE_HIGHLIGHTS = "TOGGLE_HIGHLIGHTS",
+	REMOVE_HIGHLIGHTS = "REMOVE_HIGHLIGHTS",
+	RENDER_HIGHGHLIGHTS = "RENDER_HIGHLIGHTS",
+	SELECT_COLOR = "SELECT_COLOR",
+}
+
+export interface MessageData<T> {
+	action: MessageAction;
+	data: T;
+}
+
+export enum StorageKey {
+	SHOW_HIGHLIGHTS = "SHOW_HIGHLIGHTS",
+	HIGHLIGHTING_COLOR = "HIGHLIGHTING_COLOR",
+	COLOR_SELECTION = "COLOR_SELECTION",
+	COLLAB_ID = "COLLAB_ID",
+}
