+++ src/Nostr/types.ts
@@ -6,27 +6,3 @@ export interface NostrUser {
 export interface NostrUserMetadata {
 	[pubkey: string]: string;
 }
-
-export interface ColorDescription {
-	name: string;
-	val: string;
-}
-
-export enum MessageAction {
-	TOGGLE_HIGHLIGHTS = "TOGGLE_HIGHLIGHTS",
-	REMOVE_HIGHLIGHTS = "REMOVE_HIGHLIGHTS",
-	RENDER_HIGHGHLIGHTS = "RENDER_HIGHLIGHTS",
-	SELECT_COLOR = "SELECT_COLOR",
-}
-
-export interface MessageData<T> {
-	action: MessageAction;
-	data: T;
-}
-
-export enum StorageKey {
-	SHOW_HIGHLIGHTS = "SHOW_HIGHLIGHTS",
-	HIGHLIGHTING_COLOR = "HIGHLIGHTING_COLOR",
-	COLOR_SELECTION = "COLOR_SELECTION",
-	COLLAB_ID = "COLLAB_ID",
-}
